feat: adding book inventory model and creating repository/controllers

+++ src/main/java/com/purespectrum/rest/api/repository/UserRepository.java
@@ -3,5 +3,5 @@ package com.purespectrum.rest.api.repository;
 import com.purespectrum.rest.api.model.UserModel;
 import org.springframework.data.repository.CrudRepository;
 
-public interface UserRepository extends CrudRepository<UserModel, Integer> {
+public interface UserRepository extends CrudRepository<UserModel, Long> {
 }
